AgroSisAPIDjangopast/apps/autenticacion/views.py
from django.shortcuts import render
from django.contrib.auth.models import User
from django.core.mail import send_mail
from django.utils.http import urlsafe_base64_encode
from django.utils.encoding import force_bytes
from django.contrib.auth.tokens import default_token_generator
from rest_framework.response import Response
from rest_framework.views import APIView
from rest_framework import status
from decouple import config

# para las Class PasswordReset ConfirmView
from django.utils.http import urlsafe_base64_decode
from django.utils.encoding import force_str


class PasswordResetRequestView(APIView):
    def post(self, request):
        email = request.data.get("email")
        if not email:
            return Response({"error": "El correo es requerido"}, status=status.HTTP_400_BAD_REQUEST)

        try:
            user = User.objects.get(email=email)
        except User.DoesNotExist:
            return Response({"error": "No se encontró un usuario con ese correo"}, status=status.HTTP_404_NOT_FOUND)

        uid = urlsafe_base64_encode(force_bytes(user.pk))
        token = default_token_generator.make_token(user)
        reset_link = f"{config('FRONTEND_URL')}/reset-password/{uid}/{token}/"

        send_mail(
            "Restablecer contraseña",
            f"Haz clic en el siguiente enlace para restablecer tu contraseña: {reset_link}",
            config("EMAIL_HOST_USER"),
            [email],
            fail_silently=False,
        )

        return Response({"message": "Correo enviado con el enlace para restablecer la contraseña"}, status=status.HTTP_200_OK)

class PasswordResetConfirmView(APIView):
    def post(self, request, uidb64, token):
        password = request.data.get("password")

        if not password:
            return Response({"error": "La contraseña es requerida"}, status=status.HTTP_400_BAD_REQUEST)

        try:
            uid = force_str(urlsafe_base64_decode(uidb64))
            user = User.objects.get(pk=uid)
        except (User.DoesNotExist, ValueError, TypeError):
            return Response({"error": "Enlace inválido"}, status=status.HTTP_400_BAD_REQUEST)

        if not default_token_generator.check_token(user, token):
            return Response({"error": "Token inválido o expirado"}, status=status.HTTP_400_BAD_REQUEST)

        user.set_password(password)
        user.save()

        return Response({"message": "Contraseña cambiada exitosamente"}, status=status.HTTP_200_OK)
